Massive update of PTS runner (gh-55).

+++ benchmarks/cpp/PTS/benchmark.sh
@@ -22,22 +22,30 @@ PHP="php -c $PREFIX/etc/php/8.2/cli"
 
 PTS_DIR=$HOME/src/phoronix-test-suite
 PTS_LOCAL_DIR=$HOME/.phoronix-test-suite
-
-# A list of relevant tests:
-#   - come from popular projects
-#   - no external dependencies (I don't have root access on server)
-#   - don't take too much time (e.g. ffmpeg, fftw, redis were removed)
-#   - produce stable results (e.g. all compress/XXX were removed)
-#   - respect CFLAGS/CXXFLAGS
-#     (e.g. mkl-dnn, onednn, etc. reset them and nginx uses CFLAGS for both)
-#   - total runtime of tests under 6 hours
-#     (that's free time on server that I have)
-tt='pts/apache pts/botan pts/bullet pts/coremark pts/crafty pts/c-ray pts/gcrypt pts/gmpbench pts/luajit pts/nginx pts/openssl pts/polybench-c pts/povray pts/simdjson pts/z3'
+CONFIG=$(dirname $0)/config.txt
 
 export NO_EXTERNAL_DEPENDENCIES=TRUE
 
-for t in $tt; do
-  $PHP $PTS_DIR/pts-core/phoronix-test-suite.php batch-install $t
-  setarch -R $PHP $PTS_DIR/pts-core/phoronix-test-suite.php batch-benchmark $t
-  rm -rf $PTS_LOCAL_DIR/installed-tests/$t*
-done
+sanitize() {
+  echo "$cfg" | sed -e 's/#.*//; s/^ *//; s/ *$//'
+}
+
+while read cfg; do
+  cfg=$(sanitize "$cfg")
+  test -n "$cfg" || continue
+
+  cflags=$(echo "$cfg" | awk -F: '{print $1}')
+  cxxflags=$(echo "$cfg" | awk -F: '{print $2}')
+  tt=$(echo "$cfg" | awk -F: '{print $3}')
+
+  for t in $tt; do
+    (
+      # Keep default Autoconf flags
+      export CFLAGS="${CFLAGS:--g -O2} $cflags"
+      export CXXFLAGS="${CXXFLAGS:--g -O2} $cxxflags"
+      $PHP $PTS_DIR/pts-core/phoronix-test-suite.php batch-install $t
+      setarch -R $PHP $PTS_DIR/pts-core/phoronix-test-suite.php batch-benchmark $t
+    )
+    rm -rf $PTS_LOCAL_DIR/installed-tests/$t*
+  done
+done < "$CONFIG"
